Increased performance for inverse random3d geometry
The random3d point generator now creates a KDTree (from scipy.spatial) and stores the points in there. This means double memory cost, but significant performance increases.

The drawback is that the IntervalSet is NOT supported. The KDTree does not know how many points to query, and therefore we either have to query all and select the first point inside the interval, or requery if the selected point is not inside the interval.